requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.1
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cohere==5.3.4
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.5
datasets==2.14.7
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
distro==1.9.0
dspy-ai==2.4.9
effdet==0.4.1
emoji==2.11.1
exceptiongroup==1.2.1
fastapi==0.110.3
fastavro==1.9.4
filelock==3.14.0
filetype==1.2.0
flask==3.0.3
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.10.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib-resources==6.4.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
jsonpath-python==1.0.6
kiwisolver==1.4.5
kubernetes==29.0.0
langdetect==1.0.9
layoutparser==0.3.4
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.33
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.20
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
lxml==5.2.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.25.0
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.15.1
pillow==10.3.0
pillow-heif==0.16.0
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.5.0
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
PyPika==0.48.9
pyproject-hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
striprtf==0.0.26
structlog==24.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.1
typer==0.12.3
types-requests==2.31.0.20240406
typing-extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.13.6
unstructured-client==0.22.0
unstructured-inference==0.7.26
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
